gpu: nova-core: switch to kernel bitfield macro
Use the new kernel bitfield! macro instead of the driver-local one.

zhiw: also convert mctp.rs and fsp.rs to the new bitfield API:
  - 1-bit fields use => bool, 8/16-bit fields use => u8/u16
  - sub-byte fields use with_const_* or try_with_*
  - msg_type() comparison adapted for Bounded<u32, 7> return type

+++ drivers/gpu/nova-core/fsp.rs
@@ -465,7 +465,7 @@ impl Fsp {
         }
 
         let nvdm = mctp::NvdmHeader::from_raw(nvdm_header);
-        if nvdm.msg_type() != mctp::MSG_TYPE_VENDOR_PCI
+        if u32::from(nvdm.msg_type()) != u32::from(mctp::MSG_TYPE_VENDOR_PCI)
             || nvdm.vendor_id() != mctp::VENDOR_ID_NV
             || nvdm.nvdm_type() != mctp::nvdm_type::FSP_RESPONSE
         {
